requirements.txt
Flask==3.1.0
PyNaCl==1.5.0
requests==2.32.3
